将用户ID 转换为Int

+++ common/common-lib/src/main/java/com/little/g/springcloud/common/params/UidTimeQueryParam.java
@@ -5,13 +5,13 @@ package com.little.g.springcloud.common.params;
  */
 public class UidTimeQueryParam extends TimeQueryParam {
 
-	private Long uid;
+	private Integer uid;
 
-	public Long getUid() {
+	public Integer getUid() {
 		return uid;
 	}
 
-	public void setUid(Long uid) {
+	public void setUid(Integer uid) {
 		this.uid = uid;
 	}
