+++ pay/pay-api/src/main/java/com/little/g/springcloud/pay/dto/AccountHasUserId.java
@@ -5,6 +5,6 @@ package com.little.g.springcloud.pay.dto;
  */
 public interface AccountHasUserId extends Account {
 
-	long getUid();
+    Integer getUid();
 
 }
