+++ pay/pay-api/src/main/java/com/little/g/springcloud/pay/params/PreOrderParams.java
@@ -26,10 +26,10 @@ public class PreOrderParams implements Serializable {
 	private Long totalFee;
 
 	@NotNull
-	private Long accountId;
+    private Integer accountId;
 
 	@NotNull
-	private Long oppositAccount;
+    private Integer oppositAccount;
 
 	@NotNull
 	private TradeType tradeType;
@@ -74,19 +74,19 @@ public class PreOrderParams implements Serializable {
 		this.totalFee = totalFee;
 	}
 
-	public Long getAccountId() {
+    public Integer getAccountId() {
 		return accountId;
 	}
 
-	public void setAccountId(Long accountId) {
+    public void setAccountId(Integer accountId) {
 		this.accountId = accountId;
 	}
 
-	public Long getOppositAccount() {
+    public Integer getOppositAccount() {
 		return oppositAccount;
 	}
 
-	public void setOppositAccount(Long oppositAccount) {
+    public void setOppositAccount(Integer oppositAccount) {
 		this.oppositAccount = oppositAccount;
 	}
