+++ user/user-dao/src/main/java/com/little/g/springcloud/user/mapper/UserMapperExt.java
@@ -2,6 +2,6 @@ package com.little.g.springcloud.user.mapper;
 
 public interface UserMapperExt {
 
-	Long maxUid();
+    Integer maxUid();
 
 }
